+++ firmware/src/modes/BinaryEpochMode.cpp
@@ -11,7 +11,7 @@ void BinaryEpochMode::handle()
     if (epoch != _epoch)
     {
         epoch = _epoch;
-        for (uint8_t i = 0; i < GRID_COLUMNS * GRID_ROWS / 8; ++i)
+        for (uint8_t i = 0; i < static_cast<uint8_t>(GRID_COLUMNS * GRID_ROWS / 8); ++i)
         {
             const uint8_t x = GRID_COLUMNS - 2 - (i % (GRID_COLUMNS / 2) * 2);
             const uint8_t y = GRID_ROWS - 4 - (i / (GRID_COLUMNS / 2) * 4);
